Added patch mplayer-filmdint-vfr.patch
Patch and description from:
http://lists.mplayerhq.hu/pipermail/mplayer-users/2006-October/063360.html

mplayer-filmdint-vfr.patch add a new vfr parameter to the filmdint
filter. When it is set to a non-zero value, filmdint should try to
output frames with reasonable pts values, and should try not to drop
frames unless they are telecine duplicates. This is still not perfect,
and will need either a two-pass mode or frame buffering in the filter
to fix.