Fixed and applied mplayer-tcdump.patch
Patch from http://lists.mplayerhq.hu/pipermail/mplayer-users/2006-October/063360.html
Description from manpage.

-vf tcdump[=filename]
Dump frame timecodes to a file, in the v2 external timecode file used by
mkvmerge. With proper post-processing, this could be used to produce VFR
files with mplayer/mencoder. Note that mencoder does not pass timestamps
to the filter chain, and will not work with this filter. If none is given,
the default filename is "timecodesv2.txt".