Hack to add libass compatibility for mencoder with embedded font support
Patch from http://oss.netfarm.it/mplayer/patches/sm_ass_mencoder_embedfonts.diff

This adds a new mencoder parameter, -keep-pts, that is automatically enabled by -ass. It might cause "badly interleaved frame" errors, so beware.

This also adds a new video filter (do NOT use it with -ss or EDLs):

-vf fixpts[=options]
Fixes the presentation timestamps (PTS) of the frames.
By default, the PTS passed to the next filter is dropped, but the following options can change that:

* print
Print the incoming PTS.
* fps=<fps>
Specify a frame per second value.
* start=<pts>
Specify an initial value for the PTS.
* autostart=<n>
Uses the nth incoming PTS as the initial PTS. All previous pts are kept, so setting a huge value or -1 keeps the PTS intact.
* autofps=<n>
Uses the nth incoming PTS after the end of autostart to determine the framerate.

Example:

"-vf fixpts=fps=24000/1001,ass,fixpts"
Generates a new sequence of PTS, uses it for ASS subtitles, then drops it.
Generating a new sequence is useful when the timestamps are reset during the program; this is frequent on DVDs.
Dropping it may be necessary to avoid confusing encoders.